Refact: 팀 게시판 댓글 컴포넌트 통일
